+++ test/test_contacts.dart
@@ -57,6 +57,7 @@ main() {
       ..CountryOrRegion = "Ukraine"
       ..City = "Kharkiv"
       ..Street = "Hrushevsky 23";
+    contact.ImAddresses[ImAddressKey.ImAddress1] = "1234566";
     await contact.Save();
     await contact.Delete(DeleteMode.HardDelete);
   });